feat:bye-rounds

+++ backend/tournaments/src/main/java/com/crashcourse/kickoff/tms/match/service/RoundService.java
@@ -7,5 +7,5 @@ import com.crashcourse.kickoff.tms.match.model.Round;
 import com.crashcourse.kickoff.tms.match.dto.*;
 
 public interface RoundService {
-    Round createRound(int numberOfMatches, Long roundNumber);
+    Round createRound(int numberOfMatches, int roundNumber);
 }
