+++ backend/tournaments/src/main/java/com/crashcourse/kickoff/tms/match/service/RoundServiceImpl.java
@@ -24,9 +24,9 @@ public class RoundServiceImpl implements RoundService {
     private final MatchService matchService;
 
     @Override
-    public Round createRound(int numberOfMatches, Long roundNumber) {
+    public Round createRound(int numberOfMatches, int roundNumber) {
         Round round = new Round();
-        round.setRoundNumber(roundNumber);
+        round.setRoundNumber(Long.valueOf(roundNumber));
         round = roundRepository.save(round);
     
         List<Match> matches = new ArrayList<>();
